These updates are to fix some warnings. Also remove requests from the zip files. It's fine to use it, but there is no need to put in an old version of requests as a default.

+++ AndroidApp/app/src/main/java/com/example/pythontest/PythonThread.java
@@ -24,7 +24,7 @@ public class PythonThread extends Thread
 
     // JNI links
     public native int initPython(String aPath);
-    public native int runPython(String aFilename);
+    public native long runPython(String aFilename);
     public native int cleanupPython();
 
     private final Logger mLogger;
